+++ src/reducers/index.js
@@ -1,8 +1,10 @@
 import { combineReducers } from 'redux';
 import messages from './messages';
 import users from './users';
+import currentUser from './currentUser';
 
 const chat = combineReducers({
+  currentUser,
   messages,
   users
 });