+++ src/reducers/users.js
@@ -1,13 +1,18 @@
-import { ADD_USER, USERS_LIST } from '../constants/actionsTypes';
+import { INITIAL_DATA, ADD_USER, USERS_LIST } from '../constants/actionsTypes';
 
 const users = (state = [], action) => {
   switch (action.type) {
+    case INITIAL_DATA:
+      return [
+        ...state,
+        ...action.users
+      ];
     case ADD_USER:
       return [
         ...state,
         {
           name: action.name,
-          id: action.id
+          uuid: action.uuid
         }
       ];
     case USERS_LIST: