+++ CHANGELOG.md
@@ -1,6 +1,11 @@
 ## Chat-App
 
 
+# 2.1.0
+ * Added real ID's codes for messages and users present inside chat.
+ * Improved reducers and store configs for better payload handling.
+ * Added current user reducer.
+
 # 2.0.0
  * Added ability to watch and send both users and chat messages lists for a freshly comnnected users.
  * Minor inprovements.